+++ Source/ActivationFunctions/riscv_nn_relu_any_s8.c
@@ -1,6 +1,6 @@
 /******************************************************************************
  * Copyright (C) 2010-2018 Arm Limited or its affiliates. All rights reserved.*
- * Copyright (C) 2018-2022 Andes Technology Corporation. All rights reserved. *
+ * Copyright (C) 2018-2023 Andes Technology Corporation. All rights reserved. *
  *                                                                            *
  * SPDX-License-Identifier: Apache-2.0                                        *
  *                                                                            *
@@ -24,7 +24,7 @@
 
 //// Activations Functions
 
-void riscv_nn_relu_any_s8(q7_t *data, uint16_t size, q7_t max_val)
+void riscv_nn_relu_any_s8(q7_t *data, uint32_t size, q7_t max_val)
 {
     while(size-- > 0)
     {
